requirements.txt
allennlp
allennlp-models
torchsnooper

pandas

pytest-pythonpath

# Checks style, syntax, and other useful errors.
pylint==1.8.1
